# Bytewise crc32