--- a/go.mod
+++ b/go.mod
@@ -3,7 +3,7 @@ module github.com/asticode/go-astits
 go 1.13
 
 require (
-	github.com/asticode/go-astikit v0.20.0
+	github.com/asticode/go-astikit v0.30.0
 	github.com/pkg/profile v1.4.0
 	github.com/stretchr/testify v1.4.0
 )