… crc32_table.go
… go.sum
… internal/cmd/crc32_table/main.go