Good day!

This pull request introduce bytewise crc32 calculation instead of current bitwise. Code based on VLC implementation, but instead per-object table generation, it use static crc32 table (+1kb additional lifetime memory use but no recalculations and reallocations) .
Current:
```
Benchmark_updateCRC32/Calc_PAT_crc32
Benchmark_updateCRC32/Calc_PAT_crc32-8         	 9470034	       122.9 ns/op	       0 B/op	       0 allocs/op
Benchmark_updateCRC32/Calc_PMT_crc32
Benchmark_updateCRC32/Calc_PMT_crc32-8         	 4010506	       283.4 ns/op	       0 B/op	       0 allocs/op
```
Introduced in this patch:
```
Benchmark_updateCRC32/Calc_PAT_crc32_precomputed
Benchmark_updateCRC32/Calc_PAT_crc32_precomputed-8         	71367823	        16.79 ns/op	       0 B/op	       0 allocs/op
Benchmark_updateCRC32/Calc_PMT_crc32_precomputed
Benchmark_updateCRC32/Calc_PMT_crc32_precomputed-8         	18918052	        54.75 ns/op	       0 B/op	       0 allocs/op
```

Test's and benchmarks based on real TS headers. Old implementation renamed and kept in file for comparison sake, you can remove it safely.