@asticode 2022-09-08
fyi I've created a `v1.11.0` tag

thanks for the PR! ❤️